+++ 07-iron/src/main/scala/dagmendez/iron/B_TypeAliases.scala
@@ -1,4 +1,4 @@
-package dagmendez.presentation
+package dagmendez.iron
 
 object B_TypeAliases:
