+++ 07-iron/src/main/scala/dagmendez/iron/D_ValueClassesWithValidation.scala
@@ -1,4 +1,4 @@
-package dagmendez.presentation
+package dagmendez.iron
 
 object D_ValueClassesWithValidation:
