+++ 07-iron/src/main/scala/dagmendez/iron/E_ValueClassesWithErrorHandling.scala
@@ -1,4 +1,4 @@
-package dagmendez.presentation
+package dagmendez.iron
 
 object E_ValueClassesWithErrorHandling:
