+++ 07-iron/src/main/scala/dagmendez/iron/F_OpaqueTypes.scala
@@ -1,4 +1,4 @@
-package dagmendez.presentation
+package dagmendez.iron
 
 object F_OpaqueTypes:
