+++ 07-iron/src/main/scala/dagmendez/iron/G_OpaqueTypesWithCompileError.scala
@@ -1,4 +1,4 @@
-package dagmendez.presentation
+package dagmendez.iron
 
 import scala.compiletime.constValue
 import scala.compiletime.error
